Rename ANSIOutput file
