+++ src/vs/base/common/ansiOutput.ts
