+++ src/vs/workbench/services/positronConsole/common/classes/runtimeItemStartup.ts
@@ -2,7 +2,7 @@
  *  Copyright (C) 2023 Posit Software, PBC. All rights reserved.
  *--------------------------------------------------------------------------------------------*/
 
-import { ANSIOutput, ANSIOutputLine } from 'vs/base/common/ansi-output';
+import { ANSIOutput, ANSIOutputLine } from 'vs/base/common/ansiOutput';
 import { RuntimeItem } from 'vs/workbench/services/positronConsole/common/classes/runtimeItem';
 
 /**
